fix routeS

+++ database/migrations/2022_12_21_101557_add_foreign_key_to_order_table.php
@@ -13,10 +13,12 @@ class AddForeignKeyToOrderTable extends Migration
      */
     public function up()
     {
-        Schema::table('orders', function (Blueprint $table) {
-            //
-            $table->foreign('coupon_id')->references('id')->on('coupon')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn("orders", "coupon_id")) {
+            Schema::table('orders', function (Blueprint $table) {
+                $table->foreign('coupon_id')->references('id')->on('coupon')->onDelete('cascade');
+            });
+        }
+
     }
 
     /**
